fix: resolve 22 CI mypy errors (unused type:ignore, floating->float casts, return types)
- mercury_ml.py: remove 9 unused type:ignore comments (numpy stubs now handle these)
- realworld_benchmark.py: cast np.mean() to float() for BenchmarkMetrics fields (10 arg-type errors)
- online_learning.py: cast model.predict/predict_proba returns to explicit dtypes (3 return-value errors)

+++ src/omni_mercury_engine/core/realworld_benchmark.py
@@ -607,16 +607,16 @@ class RealWorldBenchmarkRunner:
 
         # Aggregate results
         metrics = BenchmarkMetrics(
-            roc_auc=np.mean([f["roc_auc"] for f in fold_results]),
-            pr_auc=np.mean([f["pr_auc"] for f in fold_results]),
-            f1=np.mean([f["f1"] for f in fold_results]),
-            precision=np.mean([f["precision"] for f in fold_results]),
-            recall=np.mean([f["recall"] for f in fold_results]),
-            brier_score=np.mean([f["brier_score"] for f in fold_results]),
-            event_f1=np.mean([f["event_f1"] for f in fold_results]),
-            time_to_detection=np.mean([f["time_to_detection"] for f in fold_results]),
-            fit_time_ms=np.mean([f["fit_time_ms"] for f in fold_results]),
-            predict_time_ms=np.mean([f["predict_time_ms"] for f in fold_results]),
+            roc_auc=float(np.mean([f["roc_auc"] for f in fold_results])),
+            pr_auc=float(np.mean([f["pr_auc"] for f in fold_results])),
+            f1=float(np.mean([f["f1"] for f in fold_results])),
+            precision=float(np.mean([f["precision"] for f in fold_results])),
+            recall=float(np.mean([f["recall"] for f in fold_results])),
+            brier_score=float(np.mean([f["brier_score"] for f in fold_results])),
+            event_f1=float(np.mean([f["event_f1"] for f in fold_results])),
+            time_to_detection=float(np.mean([f["time_to_detection"] for f in fold_results])),
+            fit_time_ms=float(np.mean([f["fit_time_ms"] for f in fold_results])),
+            predict_time_ms=float(np.mean([f["predict_time_ms"] for f in fold_results])),
         )
 
         # Compute confidence intervals
